Add configuration for running in vscode
Add run configurations to debug the app, mage and integration tests.
Normal unit tests are runnable individually, in file and in package from
the codelens that the golang extension provides. Integration tests are
generated by sqlboiler which doesn't generate them with the correct
naming conventions (start with uppercase Test) so codelens doesn't pick
them up so instead have added a debug config for running integration
tests in a file