Get client configuration for communicating with UAA
We need to be able to configure a custom client that trusts the UAA CA
certificate. The easiest way to do this (I think) is just to rip off
Prometheus' client configuration. This should allow us to configure a
custom CA for trusting and (later) authentication such as basic auth or
mutual TLS