Refactor adding of build tags to use the Golang AST
This way we don't just insert strings and instead treat the code like
actual golang code. This isn't that useful yet but will become useful
when we want to remove/modify elements of the code generated by
SQLBoiler such as removing some tests

Ran `mage GenerateSQLBoiler` to regenerated everything and only change
was this random reordering of orgs vs. isolation segments in mysql
spaces table. I'm committing it as it is a no-op and I'm not sure if
something about parsing the AST is changing the order in which case
hopefully it is stable going forwards