Refactor DB connection creation
Rather than using sync.Once and doing global DB connection, explicitly
build controllers etc. with a *sql.DB. This makes it much easier to
inject a mocked DB rather than having to deal with the global db class.
As part of this, it makes sense to move the migrate/create code
specifically to Mage as it is a little scary having this codebase
potentially creating DBs/running connections when the primary use case
is to connect to an existing CC DB which will have it's schema
controlled in the Ruby codebase. Having local Mage commands for devs is
still helpful though.