Use SQL IN for filtering rather than combining WHEREs
Using IN is slightly more safe as then we don't have to mess around with
indexes/SQL fragments in code to get the brackets right. There is a
helper for the "name" field as it is a pure string but there is no
equivalent helper for null.String so I just wrote one and used it for
both to be consistent. We would probably move this to a helper package
eventually. Using the constants generated by sqlboiler make this a bit
safer also as we get strong typing