Use custom sqlboiler templates
We are having trouble testing sqlboiler generated code as it assumes you
will just use a mocked DB engine. This works but is awkward as our tests
for controller code become extremely tied to the internal working on
sqlboiler (e.g. you need to know how boil.Infer() will find columns).
This means it takes a long time to write the tests and they aren't that
relevant as well as being brittle.

These templates have been modified from upstream to create a
<Table>Query object (exported this time) with all the
Insert/Update/Delete etc. functions on it. Interfaces are then defined
(which <Table>Query implements) so that the consumers such as
controllers can just declare that a var is of the interface type which
can then be mocked in tests.

The autogenerated tests from sqlboiler won't work here so include the
`no-tests` flag instead