feat: CFP-1761 Implement docs with docusaurus
Adds a nodejs project called docusaurus (https://docusaurus.io/)
This generates a static page for following documentation:
- Hand written documentation
- Generated GoDocs from Go Comments
- Api Documentation from swagger file
- ADRs in the form of a blog
It supports versioned documentation as well as multi language support if
things would get serious.

The `mage generate` command will generate all the docs that can be
automatically created (godocs, swagger)
A actions workflow will check that the generateable docs are generated to the docs folder in each pr.
A actions workflow will automatically generate the page once there is a
change in the docs folder and push it as a single commit to the gh-pages
branch(history is deleted every time). This branch is served as documentation by github pages.