Generate subquerying and quote helpers
Add a quoting helper for building subqueries safely and a new Subquery
type which does not replace any placeholders or end in a ";". This
allows a subquery to be used inside a larger query and all the params
from placeholders get resolved at that point.

This required some changes to sqlboiler source code as well as templates
so rather than just running `sqlboiler`, we need to Chdir into the
sqlboiler submodule and run `go run main.go`. This requires paths in
sqlboiler config files to use a relative path.