Separate buildpack controller tests
The test file (and the controller too) was getting a bit unwieldy. We
also had a lot of duplication between test suites. This commit splits
each of the endpoints (and their relevant tests) into separate files to
hopefully make it easier to read and find code. The
controller/controller_test files exist to hold common code between all
methods/tests such as the Controller struct and the shared base suite