Use testify Mock as querier function to capture queryMods
There's an odd (potential) bug with the tests here when the querier is
called twice (e.g. in a "List" endpoint for getting the Count and All)
as only the last set of queryMods would be saved into suite.queryMods.
To avoid this potential issue and also capture every call the the
querier, we can just leverage testify's Mock type to act as a call
recorder