Apply filters to count
The "total_results" in the response should be the total number of
results for a given set of filters. We should thus apply these filters
to the Count query also. This applies those mods whilst ignoring
irrelevant ones such as OrderBy and Limit/Offset. The tests also ensure
that only the correct mods are applied to the Count.