Refactor auth token to extract scopes
The global roles (admin, admin read only, global auditor) are decided
based on the token rather than tables in the DB (like space/org roles).
This means to be able to use these roles in permissions checks we need
to extract them from the token.

By modifying the JWT success handler to persist the scopes from the
token in the echo Context, we can (with some helper methods) check if a
user is an admin etc. We can also do the upfront check that they have
the neccessary read/write scope to access a resource which is best
implemented as a middleware on a group that each route is registered
with. This should hopefully make it obvious when registering new
endpoints that they must live behind one of these groups (or go straight
on the v3 root if no read/write check is required).

To help with typing, a Scope type and some constants allow for strongly
typed references to the magic scope strings (e.g.
"cloud_controller.read") that the existing CC depends on