Calculate reset time based on offset interval
Rather than just a randomised offset since time.Now, we would like all
instances to agree on the reset time (so that headers are consistent).
As all instances should be in-sync for clocks (due to NTP), we can just
use a regular interval based on the unix timestamp for when limits
reset. To avoid the thundering herd problem, we still use an offset
which should be evenly distributed across the reset period or at least
evenly enough to mitigate the problem