Use imports to not require fork of sqlboiler
The reason that sqlboiler was forked rather than the templates just
imported was so that a) we could add extra functions and imports to some
singleton classes and b) so that we could track the changes to upstream
as a git diff. Turns out there is a way in sqlboiler to do a) and whilst
b) is a nice to have we can always do a diff on the relevant files to
see what has changed. This imports the modified templates and makes the
required changes to get the files to generate properly. The changed
sqlboiler files (some ordering changes) will be in a separate commit