Increase complexity limit in cyclop linter
The cyclop linter is designed to keep functions small and simple. The
structure of our project means that the controllers contain lots of
"dumb logic" like extracting parameters etc. This doesn't really add to
the complexity for someone reading the code but the linter thinks it is.
By increasing the complexity a bit we should still catch methods that
get far too complicated whilst also allowing for these controllers to
pass linting