Ensure that empty names= filter does not return results
The behaviour of cc_ng is such that an empty names= filter should return
no results. To replicate this behaviour we need to use a pointer in our
filter params struct so that we can tell the difference between "didn't
specify names" and "specified names but empty" and include a clause in
the query in the latter case (that should always return 0 rows as names
can't be empty)

As part of proving this works, just ported over all filter tests from
buildpacks and modified them to work for securitygroups. Changed these
tests to do everything as admin as we do permissions checks separately
in the integration tests. Changed receivers in tests from suite -> s to
avoid shadowing "suite" import package