Use correct buildpack position during post
CCNG has a special behaivor when posting (also delete) new buildpacks.
It shuffles existing buildpacks so that there is no gap in positions.
`Position` is a field in the buildpacks table which is needed for the
buildpack auto-detection.
Example:
    Current state: there are 10 buildpacks with position 1-10
    New BP with postion 1 or no position provided -> all 10 previous positions will be shifted +1
    New BP with position 1899 -> new BP will be created with last
    position plus one (11 with current state)

This commit adds the same behavior for post requests in go-cf-api.

Co-authored-by: Andrew Paine <andrew.paine@sap.com>
Co-authored-by: Johannes Haass <johannes.haass@sap.com>